# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "econpaper-backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/sessions/{session_id}/generate-chapter": {
      "post": {
        "summary": "Generate Chapter Endpoint",
        "description": "触发 generate_chapter 节点：写入 current_chapter → 跑节点 → 返回章节。\n\n返回体：\n{\n  \"chapter\": {type, title, content, status: \"generated\", ...},\n  \"body_chapters\": [...],  # 全部正文章节\n}",
        "operationId": "generate_chapter_endpoint_sessions__session_id__generate_chapter_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateChapterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateChapterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/approve-chapter": {
      "post": {
        "summary": "Approve Chapter Endpoint",
        "description": "用户审批章节 → 标记 status=\"approved\"。\n\ngraph resume 由后续集成阶段在 graph 层加 interrupt() 后实现；此处只\n更新 session 状态。",
        "operationId": "approve_chapter_endpoint_sessions__session_id__approve_chapter_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveChapterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproveChapterResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/rollback": {
      "post": {
        "summary": "Rollback Chapter Endpoint",
        "description": "回滚到指定版本。\n\nRequest:  ``{\"chapter_index\": int, \"version_index\": int}``\nResponse: ``{\"chapter\": {...}, \"body_chapters\": [...]}``\n\n把 chapter_index / version_index 写入 state 后调\n``rollback_chapter(state)`` 节点（T-08a），节点返回\n``{\"body_chapters\": [...]}``，合并回 session 状态。",
        "operationId": "rollback_chapter_endpoint_sessions__session_id__rollback_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RollbackRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RollbackResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/regenerate": {
      "post": {
        "summary": "Regenerate Chapter Endpoint",
        "description": "重新生成当前章。\n\nRequest:  ``{\"chapter_index\": int}``\nResponse: ``{\"chapter\": {...含新版本...}, \"body_chapters\": [...]}``\n\n设置 ``state['current_chapter_index']``，调 ``generate_chapter(state)``\n节点，合并结果。",
        "operationId": "regenerate_chapter_endpoint_sessions__session_id__regenerate_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegenerateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sessions/{session_id}/chapters/{chapter_index}/versions": {
      "get": {
        "summary": "Get Versions",
        "description": "获取指定章节的所有版本。\n\nResponse: ``{\"chapter_index\": int, \"count\": int,\n              \"versions\": [{\"index\": int, \"preview\": str}]}``\n\n每个版本的 ``preview`` 截断到前 50 字。章节无 ``versions`` 列表时，\n降级用 ``content`` 作为唯一版本。",
        "operationId": "get_versions_sessions__session_id__chapters__chapter_index__versions_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          },
          {
            "name": "chapter_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Chapter Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VersionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApproveChapterRequest": {
        "properties": {
          "chapter_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Type"
          }
        },
        "type": "object",
        "title": "ApproveChapterRequest",
        "description": "POST /sessions/{id}/approve-chapter 请求体。"
      },
      "ApproveChapterResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "chapter": {
            "$ref": "#/components/schemas/ChapterResponse"
          },
          "body_chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterResponse"
            },
            "type": "array",
            "title": "Body Chapters"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "chapter"
        ],
        "title": "ApproveChapterResponse",
        "description": "POST /sessions/{id}/approve-chapter 返回体。"
      },
...
      "ChapterResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": ""
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "versions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Versions"
          },
          "chapter_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Index"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "ChapterResponse",
        "description": "论文章节（generate_title / generate_chapter 写入）。\n\nstatus 枚举：``generated`` | ``approved`` | ``edited`` | ``rolled_back``。\n前端可在本地临时态再加 ``streaming`` / ``done``。"
      },
      "ChapterSpec": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "default": ""
          },
          "method": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Method"
          },
          "research_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Question"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "ChapterSpec",
        "description": "章节定义（用户选择要生成哪一章）。"
      },
      "ChapterVersionItem": {
        "properties": {
          "index": {
            "type": "integer",
            "title": "Index"
          },
          "preview": {
            "type": "string",
            "title": "Preview"
          }
        },
        "type": "object",
        "required": [
          "index",
          "preview"
        ],
        "title": "ChapterVersionItem",
        "description": "单个版本预览项。"
      },
...
      "GenerateChapterRequest": {
        "properties": {
          "chapter": {
            "$ref": "#/components/schemas/ChapterSpec"
          },
          "render_kwargs": {
            "type": "object",
            "title": "Render Kwargs"
          }
        },
        "type": "object",
        "required": [
          "chapter"
        ],
        "title": "GenerateChapterRequest",
        "description": "POST /sessions/{id}/generate-chapter 请求体。"
      },
      "GenerateChapterResponse": {
        "properties": {
          "chapter": {
            "$ref": "#/components/schemas/ChapterResponse"
          },
          "body_chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterResponse"
            },
            "type": "array",
            "title": "Body Chapters"
          }
        },
        "type": "object",
        "required": [
          "chapter"
        ],
        "title": "GenerateChapterResponse",
        "description": "POST /sessions/{id}/generate-chapter 返回体。"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RegenerateRequest": {
        "properties": {
          "chapter_index": {
            "type": "integer",
            "title": "Chapter Index",
            "default": 0
          }
        },
        "type": "object",
        "title": "RegenerateRequest",
        "description": "POST /sessions/{id}/regenerate 请求体。"
      },
      "RegenerateResponse": {
        "properties": {
          "chapter": {
            "$ref": "#/components/schemas/ChapterResponse"
          },
          "body_chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterResponse"
            },
            "type": "array",
            "title": "Body Chapters"
          }
        },
        "type": "object",
        "required": [
          "chapter"
        ],
        "title": "RegenerateResponse",
        "description": "POST /sessions/{id}/regenerate 返回体。"
      },
...
      "RollbackRequest": {
        "properties": {
          "chapter_index": {
            "type": "integer",
            "title": "Chapter Index",
            "default": 0
          },
          "version_index": {
            "type": "integer",
            "title": "Version Index",
            "default": 0
          }
        },
        "type": "object",
        "title": "RollbackRequest",
        "description": "POST /sessions/{id}/rollback 请求体。"
      },
      "RollbackResponse": {
        "properties": {
          "chapter": {
            "$ref": "#/components/schemas/ChapterResponse"
          },
          "body_chapters": {
            "items": {
              "$ref": "#/components/schemas/ChapterResponse"
            },
            "type": "array",
            "title": "Body Chapters"
          }
        },
        "type": "object",
        "required": [
          "chapter"
        ],
        "title": "RollbackResponse",
        "description": "POST /sessions/{id}/rollback 返回体。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VersionsResponse": {
        "properties": {
          "chapter_index": {
            "type": "integer",
            "title": "Chapter Index"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ChapterVersionItem"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "chapter_index",
          "count"
        ],
        "title": "VersionsResponse",
        "description": "GET /sessions/{id}/chapters/{chapter_index}/versions 返回体。"
      }
    }
  }
}